requirements.txt
langchain
langchain-community
langchain-huggingface
langchain-text-splitters
faiss-cpu
pypdf
python-dotenv
sentence-transformers
transformers
torch
streamlit


